React/add copy button component (#929)
* initial commits

* add button copy

* add styles

* refactor to use ButtonCopy

* refactor to use ButtonCopy

* refactor to use ButtonCopy

* refactor to use ButtonCopy

* refactor styles

* render icon for backstop

* add new screenshot for buttonCopy

* remove doc

* export buttonCopy

* add changelog
