+++ react/backstop/storyBookBackstop.js
@@ -100,7 +100,8 @@ const mapComponents = (components, debug) => components.map((component) => {
   // Backstop overrides.
   const overrides = [
     'GeneralTeaser',
-    'Icon'
+    'Icon',
+    'ButtonCopy'
   ];
   const backstop = (overrides.indexOf(name) > -1) ? '&backstop=true' : '';
   const url = `${urlBase}iframe.html?id=${toId(kind, name)}${backstop}`;
