+++ react/src/components/atoms/buttons/ButtonCopy/styles.scss
@@ -1,0 +1,6 @@
+@import 'global';
+
+.ma__button-copy {
+  margin-left: 0.5em;
+  @include ma_button_reset;
+}
