+++ react/src/components/base/Icon/IconDisplay.js
@@ -1,35 +1,18 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import Icon from './index';
+import ButtonCopy from '../../atoms/buttons/ButtonCopy';
 import './display.css';
 
 const IconDisplay = (props) => {
   const { name } = props;
-  const [copied, setCopied] = useState(false);
-  const copyAction = () => {
-    setCopied(true);
-    navigator.clipboard.writeText(name);
-    const timer = setTimeout(() => {
-      setCopied(false);
-    }, 1000);
-    return() => clearTimeout(timer);
-  };
-  const copyButtonTitle = copied ? 'copied' : 'copy hex code';
   return(
     <li style={{ width: 180, margin: 10, padding: 10 }} >
       <div className="sg-icons-info">
         <Icon {...props} />
         <div>
           <span>{name}</span>
-          { navigator && navigator.clipboard && (
-            <button
-              onClick={copyAction}
-              title={copyButtonTitle}
-              aria-label={copyButtonTitle}
-            >
-              { copied ? <Icon name="inputsuccess" svgWidth={16} svgHeight={16} /> : <Icon name="copy" svgWidth={16} svgHeight={16} />}
-            </button>
-          )}
+          <ButtonCopy content={name} />
         </div>
       </div>
     </li>
