+++ react/src/components/base/Typography/styles.scss
@@ -1,4 +1,4 @@
-@import '../../styles/global';
+@import 'global';
 @import "01-atoms/global/fonts-styles";
 
 .sb {
